Implemented commandline argument parsing and config group
Configuration is read from JSON data, which can either be
constructed from environmental variables, or read from a
file. This configuration is validated against
src/fattoush/config/file_input_schema.json before being
used to create a set of config objects, which are run.

All that is left here is to implement the config class and
the hooks to be run once control is passed to lettuce.

---------1---------2---------3---------4---------5---------6